+++ src/main/java/io/github/notenoughmail/tfcgenviewer/api/cache/ClimateFeatureCache.java
@@ -3,6 +3,7 @@ package io.github.notenoughmail.tfcgenviewer.api.cache;
 import com.google.common.base.Suppliers;
 import com.mojang.serialization.Codec;
 import io.github.notenoughmail.tfcgenviewer.TFCGenViewer;
+import io.github.notenoughmail.tfcgenviewer.api.SerializationInformation;
 import io.github.notenoughmail.tfcgenviewer.api.SynchronizationRequest;
 import io.github.notenoughmail.tfcgenviewer.api.color.ColorDefinition;
 import io.github.notenoughmail.tfcgenviewer.api.color.Colors;
@@ -13,7 +14,6 @@ import net.dries007.tfc.util.data.DataManager;
 import net.dries007.tfc.world.placement.ClimatePlacement;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderSet;
-import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
@@ -74,14 +74,9 @@ public class ClimateFeatureCache<C> {
                     BIOME_MOB_SETTINGS_UNIT.get()
             ), Biome::getGenerationSettings);
 
-    @Nullable
-    public static <T> Codec<T> codecForRegistry(ResourceKey<? extends Registry<T>> registry) {
-        if (Registries.PLACED_FEATURE.equals(registry)) {
-            return TFCGenViewer.cast(MINIMAL_FEATURE_CODEC);
-        } else if (Registries.BIOME.equals(registry)) {
-            return TFCGenViewer.cast(MINIMAL_BIOME_CODEC);
-        }
-        return null;
+    public static void codecForRegistry(SerializationInformation serializationInformation) {
+        serializationInformation.provide(Registries.PLACED_FEATURE, MINIMAL_FEATURE_CODEC);
+        serializationInformation.provide(Registries.BIOME, MINIMAL_BIOME_CODEC);
     }
 
     public static void syncRequest(SynchronizationRequest synchronizationRequest, @Nullable Predicate<Holder<Biome>> biomeFilter) {