+++ src/main/java/io/github/notenoughmail/tfcgenviewer/impl/visualizers/chunk/ChunkClimateRestrictedVisualizer.java
@@ -1,8 +1,8 @@
 package io.github.notenoughmail.tfcgenviewer.impl.visualizers.chunk;
 
-import com.mojang.serialization.Codec;
 import io.github.notenoughmail.tfcgenviewer.api.DrawParallelism;
 import io.github.notenoughmail.tfcgenviewer.api.MutableImage;
+import io.github.notenoughmail.tfcgenviewer.api.SerializationInformation;
 import io.github.notenoughmail.tfcgenviewer.api.SynchronizationRequest;
 import io.github.notenoughmail.tfcgenviewer.api.cache.ChunkDataProvider;
 import io.github.notenoughmail.tfcgenviewer.api.cache.ClimateFeatureCache;
@@ -15,12 +15,9 @@ import io.github.notenoughmail.tfcgenviewer.impl.TFCGenViewerRegistration;
 import net.dries007.tfc.world.TFCChunkGenerator;
 import net.dries007.tfc.world.biome.BiomeExtension;
 import net.dries007.tfc.world.chunkdata.ChunkData;
-import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.resources.ResourceKey;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.Iterator;
 import java.util.List;
@@ -124,9 +121,8 @@ public class ChunkClimateRestrictedVisualizer implements ITFCChunkVisualizerType
         ClimateFeatureCache.syncRequest(synchronizationRequest, null);
     }
 
-    @Nullable
     @Override
-    public <T> Codec<T> elementCodecForRegistry(ResourceKey<? extends Registry<T>> registry) {
-        return ClimateFeatureCache.codecForRegistry(registry);
+    public void elementCodecForRegistry(SerializationInformation serializationInformation) {
+        ClimateFeatureCache.codecForRegistry(serializationInformation);
     }
 }